+++ CHANGES.rst
@@ -4,8 +4,9 @@ Changes
 dev
 ---
 
-- django 3.0 compatibility
+- django 3.0+ compatibility
 - support for django < 2.2 dropped
+- use py.test instead of nose in tests
 
 1.4 (30-05-2017)
 ----------------